requirements.txt
torch>=1.7.0
torchvision>=0.8.0
numpy>=1.19.0
pandas>=1.1.0
scikit-learn>=0.23.0
matplotlib>=3.3.0
tqdm>=4.50.0
seaborn>=0.11.0
biopython>=1.78
scipy>=1.5.0
joblib>=0.17.0 